Move url creation code into own files.
