+++ app/javascript/availability/components/special_request_link.jsx
@@ -1,6 +1,8 @@
 import PropTypes from 'prop-types';
 import { useEffect, useState } from 'react';
 import availability from '../index';
+import illLinkUrl from '../ill_link_url';
+import aeonLinkUrl from '../aeon_link_url';
 import SpinnerLink from './spinner_link';
 
 const SpecialRequestLink = ({ holding, locationText }) => {
@@ -28,10 +30,21 @@ const SpecialRequestLink = ({ holding, locationText }) => {
           const title = encodeURIComponent(data.title_245ab_tsim);
           const author = encodeURIComponent(data.author_tsim || '');
           const pubDate = data.pub_date_illiad_ssm || '';
+          const urlArgs = {
+            linkUrl,
+            data,
+            title,
+            author,
+            pubDate,
+            holding,
+            callNumber,
+          };
           if (locationText) {
-            linkUrl = aeonLinkUrl(linkUrl, data, title, author, pubDate);
+            urlArgs.locationText = locationText;
+            urlArgs.catkey = catkey;
+            linkUrl = aeonLinkUrl(urlArgs);
           } else {
-            linkUrl = illLinkUrl(linkUrl, data, title, author, pubDate);
+            linkUrl = illLinkUrl(urlArgs);
           }
         }
       })
@@ -41,80 +54,6 @@ const SpecialRequestLink = ({ holding, locationText }) => {
         setUrl(linkUrl);
       });
   };
-
-  const illLinkUrl = (linkUrl, data, title, author, pubDate) => {
-    const itemLocation = encodeURIComponent(holding.locationID);
-    linkUrl += 'upm/illiad.dll/OpenURL?Action=10';
-    if (linkType === 'archival-thesis') {
-      linkUrl += '&Form=20&Genre=GenericRequestThesisDigitization';
-    } else {
-      const ISBN = data.isbn_valid_ssm[0] ? data.isbn_valid_ssm[0] : '';
-      linkUrl += `&Form=30&isbn=${ISBN}`;
-    }
-    if (linkType === 'reserves-scan') {
-      linkUrl += `&Genre=GenericRequestReserves&location=${itemLocation}`;
-    }
-    if (linkType === 'news-microform-scan') {
-      linkUrl += `&Genre=GenericRequestMicroScan&location=${itemLocation}`;
-    }
-    linkUrl += `&title=${title}&callno=${callNumber}&rfr_id=info%3Asid%2Fcatalog.libraries.psu.edu`;
-    if (author) {
-      linkUrl += `&aulast=${author}`;
-    }
-    if (pubDate) {
-      linkUrl += `&date=${pubDate}`;
-    }
-    return linkUrl;
-  };
-
-  const illLinkType = () => {
-    if (availability.isReserves(holding)) return 'reserves-scan';
-    if (availability.isMicroform(holding)) return 'news-microform-scan';
-    if (availability.isArchivalThesis(holding)) return 'archival-thesis';
-
-    return 'request-via-ill';
-  };
-
-  const linkType = encodeURIComponent(illLinkType());
-
-  const aeonLinkUrl = (linkUrl, data, title, author, pubDate) => {
-    const itemLocation = encodeURIComponent(locationText);
-    const itemID = encodeURIComponent(holding.itemID);
-    const { itemTypeID } = holding;
-    const genre = itemTypeID === 'ARCHIVES' ? 'ARCHIVES' : 'BOOK';
-    linkUrl =
-      'https://aeon.libraries.psu.edu/Logon/?Action=10&Form=30' +
-      `&ReferenceNumber=${catkey}&Genre=${genre}&Location=${itemLocation}` +
-      `&ItemNumber=${itemID}&CallNumber=${callNumber}` +
-      `&ItemTitle=${title}&ItemAuthor=${author}&ItemEdition=${edition(data)}` +
-      `&ItemPublisher=${publisher(data)}&ItemPlace=${pubPlace(data)}` +
-      `&ItemDate=${pubDate}&ItemInfo1=${restrictions(data)}` +
-      `&SubLocation=${subLocation(data)}`;
-    return linkUrl;
-  };
-
-  const publisher = (data) =>
-    encodeURIComponent(data.publisher_name_ssm ? data.publisher_name_ssm : '');
-
-  const pubPlace = (data) =>
-    encodeURIComponent(
-      data.publication_place_ssm ? data.publication_place_ssm : ''
-    );
-
-  const edition = (data) =>
-    encodeURIComponent(
-      data.edition_display_ssm ? data.edition_display_ssm : ''
-    );
-
-  const restrictions = (data) =>
-    encodeURIComponent(
-      data.restrictions_access_note_ssm ? data.restrictions_access_note_ssm : ''
-    );
-
-  const subLocation = (data) =>
-    encodeURIComponent(
-      data.sublocation_ssm ? data.sublocation_ssm.join('; ') : ''
-    );
 
   const label = () => {
     let text = '';
